文件下载进度使用Float

+++ HttpUtils/src/main/java/com/zhangteng/httputils/result/coroutine/callback/DeferredDownloadCallBack.kt
@@ -35,7 +35,7 @@ abstract class DeferredDownloadCallBack(
                         override fun onResponseProgress(
                             bytesRead: Long,
                             contentLength: Long,
-                            progress: Int,
+                            progress: Float,
                             done: Boolean,
                             filePath: String?
                         ) {
