+++ HttpUtils/src/main/java/com/zhangteng/httputils/result/flow/callback/FlowDownloadCallBack.kt
@@ -40,7 +40,7 @@ abstract class FlowDownloadCallBack(
                         override fun onResponseProgress(
                             bytesRead: Long,
                             contentLength: Long,
-                            progress: Int,
+                            progress: Float,
                             done: Boolean,
                             filePath: String?
                         ) {
