+++ HttpUtils/src/main/java/com/zhangteng/httputils/result/rxjava/observer/DownloadObserver.kt
@@ -63,7 +63,7 @@ abstract class DownloadObserver(
                             override fun onResponseProgress(
                                 bytesRead: Long,
                                 contentLength: Long,
-                                progress: Int,
+                                progress: Float,
                                 done: Boolean,
                                 filePath: String?
                             ) {
